worked on my logout route

+++ frontend/src/components/ArticleList.jsx
@@ -1,22 +1,31 @@
 import React from "react";
 import APIservice from "../APIService";
+import { useCookies } from 'react-cookie';
 
-function ArticleList(props) {
+function ArticleList(props) { 
+  
+  const [token] = useCookies(['mytoken']);
+
   const editBtn = (article) => {
-    // Call the editBtn function passed as a prop, and pass the clicked article as an argument
+    // Call the editBtn function passed as a prop and pass the clicked article as an argument
     props.editBtn(article);
   };
+
   const deleteBtn = (article) => {
-    APIservice.deleteArticle(article.id)
-      .then(() => props.deleteBtn(article))
-      .catch((error) => {
-        // Handle error (optional)
-        console.error('Error deleting article:', error);
-        // Add any error handling you want here.
-      });
+    // Show a confirmation dialog before proceeding with the deletion
+    const isConfirmed = window.confirm("Are you sure you want to delete this article?");
+
+    if (isConfirmed) {
+      APIservice.deleteArticle(article.id, token['mytoken'])
+        .then(() => props.deleteBtn(article))
+        .catch((error) => {
+          // Handle error (optional)
+          console.error('Error deleting article:', error);
+          // Add any error handling you want here.
+        });
+    }
   };
   
-
   return (
     <div>
       {props.articles &&
